listing new dataset for restaurants
listing 7k active restaurant businesses currently registered with the Office of Finance in LA. We can consider using this dataset instead of yelp since it includes more restaurant data and might be more interesting for our analysis. For example, is there a correlation between the total number of restaurants in a given district and the crime rate/citation? FYI, this dataset has not yet been grouped by districts.